+++ CHANGELOG.md
@@ -2,6 +2,8 @@ master
 ------
 
 * Drop support for Ruby `< 2.1.0` and Rails `4.0.0, < 3.2.0`
+* Introduce `rails g ember-cli:heroku` generator to configure a project for
+  deploying to Heroku
 * Introduce `include_ember_index_html` helper
 
 0.3.5